Move utils and imports to prevent direct import from SimulationTimeSeries
- Move definition of ProviderSet til webviz_subsurface/_utils
- Move provider set create methods and rename file
webviz_subsurface/_utils/provider_set_factory.py
- Remove webvis_subsurface/_utils/datetime_utils.py as it should not be
a common util. Due to limitation of date data frequency of daily it can be
misused and one can loose date content. It should be handled
specific in plugins.
- Update imports accordingly
- Update unit test file placement accordingly
