+++ tests/unit_tests/plugin_tests/test_simulation_time_series/test_utils/test_delta_ensemble_utils.py
@@ -2,6 +2,8 @@ from typing import List
 
 import pytest
 
+from webviz_subsurface._utils.provider_set import ProviderSet
+
 # pylint: disable=line-too-long
 from webviz_subsurface.plugins._simulation_time_series._views._subplot_view._utils.delta_ensemble_utils import (
     DeltaEnsemble,
@@ -12,12 +14,7 @@ from webviz_subsurface.plugins._simulation_time_series._views._subplot_view._uti
     is_delta_ensemble_providers_in_provider_set,
 )
 
-# pylint: disable=line-too-long
-from webviz_subsurface.plugins._simulation_time_series._views._subplot_view._utils.provider_set import (
-    ProviderSet,
-)
-
-from ..mocks.ensemble_summary_provider_dummy import EnsembleSummaryProviderDummy
+from ....mocks.ensemble_summary_provider_dummy import EnsembleSummaryProviderDummy
 
 # *******************************************************************
 #####################################################################
