+++ tests/unit_tests/plugin_tests/test_simulation_time_series/test_utils/test_derived_ensemble_vectors_accessor_utils.py
@@ -2,6 +2,7 @@ from typing import Dict, List
 
 import pandas as pd
 
+from webviz_subsurface._utils.provider_set import ProviderSet
 from webviz_subsurface.plugins._simulation_time_series._views._subplot_view._types import (
     DeltaEnsemble,
 )
@@ -22,9 +23,6 @@ from webviz_subsurface.plugins._simulation_time_series._views._subplot_view._uti
 )
 from webviz_subsurface.plugins._simulation_time_series._views._subplot_view._utils.derived_vectors_accessor.derived_vectors_accessor import (
     DerivedVectorsAccessor,
-)
-from webviz_subsurface.plugins._simulation_time_series._views._subplot_view._utils.provider_set import (
-    ProviderSet,
 )
 
 from ..mocks.derived_vectors_accessor_ensemble_summary_provider_mock import (
