+++ tests/unit_tests/plugin_tests/test_simulation_time_series/test_utils/test_history_vectors.py
@@ -13,7 +13,7 @@ from webviz_subsurface.plugins._simulation_time_series._views._subplot_view._uti
     create_history_vectors_df,
 )
 
-from ..mocks.ensemble_summary_provider_dummy import EnsembleSummaryProviderDummy
+from ....mocks.ensemble_summary_provider_dummy import EnsembleSummaryProviderDummy
 
 # *******************************************************************
 #####################################################################
