+++ tests/unit_tests/utils_tests/test_provider_set/mocks/provider_set_tests_ensemble_summary_provider_mock.py
@@ -2,7 +2,7 @@ from typing import Dict, List, Optional
 
 from webviz_subsurface._providers import VectorMetadata
 
-from ..mocks.ensemble_summary_provider_dummy import EnsembleSummaryProviderDummy
+from ....mocks.ensemble_summary_provider_dummy import EnsembleSummaryProviderDummy
 
 ##################################################################################################
 #
