+++ webviz_subsurface/_utils/provider_set_factory.py
@@ -7,7 +7,7 @@ from webviz_subsurface._providers import (
     Frequency,
 )
 
-from .._views._subplot_view._utils.provider_set import ProviderSet
+from .provider_set import ProviderSet
 
 
 def create_lazy_provider_set_from_paths(
