+++ webviz_subsurface/plugins/_prod_misfit/_plugin.py
@@ -7,14 +7,10 @@ from webviz_config import WebvizPluginABC, WebvizSettings
 
 from webviz_subsurface._models import WellAttributesModel
 from webviz_subsurface._providers import Frequency
-
-from .._simulation_time_series._utils.create_provider_set_from_paths import (
+from webviz_subsurface._utils.provider_set_factory import (
     create_presampled_provider_set_from_paths,
 )
 
-# from .._simulation_time_series.types.provider_set import (
-#     create_presampled_provider_set_from_paths,
-# )
 from ._plugin_ids import PluginIds
 from .shared_settings import Filter
 from .views import (
