+++ webviz_subsurface/plugins/_prod_misfit/utils/make_dataframes.py
@@ -4,9 +4,7 @@ from typing import Dict, List
 
 import pandas as pd
 
-from ..._simulation_time_series._views._subplot_view._utils.provider_set import (
-    ProviderSet,
-)
+from webviz_subsurface._utils.provider_set import ProviderSet
 
 
 # -------------------
