+++ webviz_subsurface/plugins/_prod_misfit/views/_well_production_coverage.py
@@ -5,11 +5,8 @@ from dash import Input, Output, callback
 from dash.development.base_component import Component
 from webviz_config.webviz_plugin_subclasses import SettingsGroupABC, ViewABC
 
-from ..._simulation_time_series._views._subplot_view._utils.provider_set import (
-    ProviderSet,
-)
+from webviz_subsurface._utils.provider_set import ProviderSet
 
-# from ..._simulation_time_series.types.provider_set import ProviderSet
 from .._plugin_ids import PluginIds
 from ..utils import make_dataframes as makedf
 from ..utils import make_figures as makefigs
