+++ webviz_subsurface/plugins/_simulation_time_series/_views/_subplot_view/_settings/_resampling_frequency.py
@@ -7,9 +7,9 @@ from webviz_config.utils import StrEnum
 from webviz_config.webviz_plugin_subclasses import SettingsGroupABC
 
 from webviz_subsurface._providers import Frequency
+from webviz_subsurface._utils.provider_set import ProviderSet
 
 from .._utils import datetime_utils
-from .._utils.provider_set import ProviderSet
 
 
 class ResamplingFrequencySettings(SettingsGroupABC):
