+++ webviz_subsurface/plugins/_simulation_time_series/_views/_subplot_view/_utils/__init__.py
@@ -3,4 +3,3 @@ from .derived_vectors_accessor import (
     DerivedEnsembleVectorsAccessorImpl,
     DerivedVectorsAccessor,
 )
-from .provider_set import ProviderSet
